requirements.txt
pandas==2.0.0
numpy==1.24.3
openai==1.3.0
streamlit==1.28.0
streamlit-chat==0.1.1
python-dotenv==1.0.0 